+++ src/PatternMatching/Parser/index.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * pattern matching parser functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/_Tag.php';
+require_once __DIR__ . '/_Tokenize.php';
+require_once __DIR__ . '/_Tokens.php';
+require_once __DIR__ . '/_Type.php';