+++ tests/Functional/RejectTest.php
@@ -14,6 +14,7 @@ class RejectTest extends \PHPUnit\Framework\TestCase
           return $val % 2 == 0;
         },
         \range(1, 5),
+        0,
         [0 => 1, 2 => 3, 4 => 5],
       ],
       [
@@ -25,7 +26,96 @@ class RejectTest extends \PHPUnit\Framework\TestCase
           ) > 3;
         },
         (object) ['foo', 'bar', 'foo-bar'],
+        0,
         (object) [0 => 'foo', 1 => 'bar'],
+      ],
+      [
+        function ($val, $key) {
+          return (
+            (
+              \extension_loaded('mbstring') ?
+                '\mb_strlen' :
+                '\strlen'
+            )($key) > 3
+          ) && f\equals($val % 2, 0);
+        },
+        [
+          'foo'  => 12,
+          'fooz' => 4,
+          'bar'  => 223,
+          'baz'  => 25,
+        ],
+        ARRAY_FILTER_USE_BOTH,
+        [
+          'foo' => 12,
+          'bar' => 223,
+          'baz' => 25,
+        ],
+      ],
+      [
+        function ($val, $key) {
+          return (
+            (
+              \extension_loaded('mbstring') ?
+                '\mb_strlen' :
+                '\strlen'
+            )($key) > 3
+          ) && f\equals($val % 2, 0);
+        },
+        (object) [
+          'foo'  => 12,
+          'fooz' => 4,
+          'bar'  => 223,
+          'baz'  => 25,
+        ],
+        ARRAY_FILTER_USE_BOTH,
+        (object) [
+          'foo' => 12,
+          'bar' => 223,
+          'baz' => 25,
+        ],
+      ],
+      [
+        function (string $key) {
+          return (
+            \extension_loaded('mbstring') ?
+              '\mb_strlen' :
+              '\strlen'
+          )($key) > 3;
+        },
+        (object) [
+          'foo'  => 12,
+          'fooz' => 4,
+          'bar'  => 223,
+          'baz'  => 25,
+        ],
+        ARRAY_FILTER_USE_KEY,
+        (object) [
+          'foo' => 12,
+          'bar' => 223,
+          'baz' => 25,
+        ],
+      ],
+      [
+        function (string $key) {
+          return (
+            \extension_loaded('mbstring') ?
+              '\mb_strlen' :
+              '\strlen'
+          )($key) > 3;
+        },
+        [
+          'foo'  => 12,
+          'fooz' => 4,
+          'bar'  => 223,
+          'baz'  => 25,
+        ],
+        ARRAY_FILTER_USE_KEY,
+        [
+          'foo' => 12,
+          'bar' => 223,
+          'baz' => 25,
+        ],
       ],
     ];
   }
@@ -33,9 +123,9 @@ class RejectTest extends \PHPUnit\Framework\TestCase
   /**
    * @dataProvider contextProvider
    */
-  public function testRejectRemovesItemsThatConformToBooleanPredicate($func, $list, $res)
+  public function testRejectRemovesItemsThatConformToBooleanPredicate($func, $list, $mode, $res)
   {
-    $filter = f\reject($func, $list);
+    $filter = f\reject($func, $list, $mode);
 
     $this->assertEquals($res, $filter);
   }