+++ tests/PatternMatching/MatchTest.php
@@ -46,13 +46,14 @@ class MatchTest extends \PHPUnit\Framework\TestCase
    */
   public function testcmatchComputesExactMatches($patterns, $entries, $res)
   {
-    $eval                 = p\cmatch($patterns);
-    [$fst, $snd, $thd]    = $entries;
-    [$rfst, $rsnd, $rthd] = $res;
+    $eval   = p\cmatch($patterns);
+    $result = [];
 
-    $this->assertEquals($rfst, $eval($fst));
-    $this->assertEquals($rsnd, $eval($snd));
-    $this->assertEquals($rthd, $eval($thd));
+    foreach ($entries as $entry) {
+      $result[] = $eval($entry);
+    }
+
+    $this->assertEquals($res, $result);
   }
 
   public function patternMatchProvider()
@@ -60,13 +61,13 @@ class MatchTest extends \PHPUnit\Framework\TestCase
     return [
       [
         [
-          '["hello", name]' => function ($name) {
+          '["hello", name]'           => function ($name) {
             return f\concat(' ', 'Hello', $name);
           },
-          IO::class => function () {
+          \sprintf('"%s"', IO::class) => function () {
             return 'i/o';
           },
-          '_' => function () {
+          '_'                         => function () {
             return 'undefined';
           },
         ],
@@ -90,13 +91,13 @@ class MatchTest extends \PHPUnit\Framework\TestCase
       ],
       [
         [
-          \stdClass::class => function () {
+          \sprintf('"%s"', \stdClass::class)  => function () {
             return 'std';
           },
-          IO::class => function () {
+          \sprintf('"%s"', IO::class)         => function () {
             return 'i/o';
           },
-          '_' => function () {
+          '_'                                 => function () {
             return 'undefined';
           },
         ],
@@ -105,28 +106,54 @@ class MatchTest extends \PHPUnit\Framework\TestCase
       ],
       [
         [
-          '[_, "bar"]' => function () {
+          '[_, "bar"]'        => function () {
             return 2;
           },
-          '["get", name]' => function ($name) {
+          '["get", name]'     => function ($name) {
             return f\concat(' ', 'Hello', $name);
           },
-          '["foo", "baz"]' => function () {
+          '["foo", 12]'       => function () {
             return 'foo-bar';
           },
-          '[a, (x:xs), b]' => function () {
-            return 12;
+          '[a, (x:xs), b]'    => function (...$args) {
+            return f\mean(f\flatten($args));
           },
-          '_' => function () {
+          '[12.224, _]'       => function () {
+            return 'float';
+          },
+          '[_, name, "foo"]'  => function ($name) {
+            return \sprintf('Hello, %s', $name);
+          },
+          '(a:b:c:d:e:_)'     => function (...$args) {
+            return f\mean($args);
+          },
+          'a'                 => function ($value) {
+            return \strtoupper($value);
+          },
+          '_'                 => function () {
             return 'undefined';
           },
         ],
         [
           ['get', 'World'],
-          [3, ['foo', 'bar'], 'baz'],
+          [3, [5, 8], 12],
           ['xyz'],
+          [12, 'Loki', 'foo'],
+          [12.224, 'baz'],
+          ['foo', 12],
+          'string',
+          \range(1, 5),
         ],
-        ['Hello World', 12, 'undefined'],
+        [
+          'Hello World',
+          7,
+          'undefined',
+          'Hello, Loki',
+          'float',
+          'foo-bar',
+          'STRING',
+          3,
+        ],
       ],
     ];
   }
@@ -136,13 +163,14 @@ class MatchTest extends \PHPUnit\Framework\TestCase
    */
   public function testpatternMatchPerformsExhaustiveMatch($patterns, $entries, $res)
   {
-    $eval                 = f\partial(p\patternMatch, $patterns);
-    [$fst, $snd, $thd]    = $entries;
-    [$rfst, $rsnd, $rthd] = $res;
+    $eval   = f\partial(p\patternMatch, $patterns);
+    $result = null;
 
-    $this->assertEquals($rfst, $eval($fst));
-    $this->assertEquals($rsnd, $eval($snd));
-    $this->assertEquals($rthd, $eval($thd));
+    foreach ($entries as $entry) {
+      $result[] = $eval($entry);
+    }
+
+    $this->assertEquals($res, $result);
   }
 
   public function letInProvider()
@@ -182,5 +210,54 @@ class MatchTest extends \PHPUnit\Framework\TestCase
     $let = p\letIn($pattern, $list);
 
     $this->assertEquals($res, $let($arg, $func));
+  }
+
+  public static function extractProvider(): iterable
+  {
+    return [
+      [
+        '["foo", _, (x:xs)]',
+        ['foo', null, \range(1, 3)],
+        [
+          'x'  => 1,
+          'xs' => [1 => 2, 2 => 3],
+        ],
+      ],
+      [
+        '(x:xs:_)',
+        \range(1, 5),
+        [
+          'x'   => 1,
+          'xs'  => 2,
+        ],
+      ],
+      [
+        '[12.23, 2, _, a]',
+        [12.23, 2, 'bar'],
+        [],
+      ],
+      [
+        'x',
+        12,
+        ['x' => 12],
+      ],
+      [
+        'x',
+        [24.2, ['foo', 'bar']],
+        [
+          'x' => [24.2, ['foo', 'bar']],
+        ],
+      ],
+    ];
+  }
+
+  /**
+   * @dataProvider extractProvider
+   */
+  public function testextractEffectsPatternBasedDestructuring(string $pattern, $input, $result)
+  {
+    $extracts = p\extract($pattern, $input);
+
+    $this->assertEquals($result, $extracts);
   }
 }